refactor: move file upload constants to lib
