+++ apps/studio/src/hooks/useLinkEditorFileMetaSuffix/utils/toDisplayType.ts
@@ -1,4 +1,4 @@
-import { FILE_UPLOAD_ACCEPTED_MIME_TYPE_MAPPING } from "~/features/editing-experience/components/form-builder/renderers/controls/constants"
+import { FILE_UPLOAD_ACCEPTED_MIME_TYPE_MAPPING } from "~/lib/fileUpload"
 
 /**
  * Converts a file extension to a display type.
