+++ apps/studio/src/utils/__tests__/formatFileSizeLimit.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from "vitest"
-import { ONE_MB_IN_BYTES } from "~/features/editing-experience/components/form-builder/renderers/controls/constants"
-
+import { ONE_MB_IN_BYTES } from "~/lib/fileUpload"
 import { formatFileSizeLimit } from "../formatFileSizeLimit"
 
 describe("formatFileSizeLimit", () => {
